+++ cmd/tomagnet/main.go
@@ -16,7 +16,7 @@ import (
 )
 
 var (
-	version = "0.3.10"
+	version = "0.3.11"
 	commit  = ""
 	date    = ""
 )